+++ code/+zapit/+gui/+main/@controller/zapSite_Callback.m
@@ -5,8 +5,6 @@ function zapSite_Callback(obj,~,~)
         val = obj.TestSiteDropDown.Value;
         f = find(cellfun(@(x) strcmp(x,val), obj.TestSiteDropDown.Items));
 
-        obj.model.stimConfig.makeChanSamples(10); % TOOD-- hard-coded LASER POWER!!
-
         % TODO -- this old structure! we must change it
         newTrial.area = f; % first brain area on the list
         newTrial.LaserOn = 1;